ceph-ansible: add post script to deploy ceph

+++ ceph-ansible/run-ansible.sh
@@ -1,16 +1,26 @@
 #!/bin/bash
 
-workdir=$(cd $(dirname $0) && pwd)
+LOCKFILE=/tmp/ceph-lock.txt
+if [ -e ${LOCKFILE} ] && kill -0 `cat ${LOCKFILE}`; then
+    echo "already running"
+    exit
+fi
 
-cd $workdir
+# make sure the lockfile is removed when we exit and then claim it
+trap "rm -f ${LOCKFILE}; exit" INT TERM EXIT
+echo $$ > ${LOCKFILE}
+
+workdir=$(cd $(dirname $0) && pwd)
 
 # install base packages
 if [[ "$(hostname)" = "ceph1" ]]; then
+    cd $workdir
+    export https_proxy=10.211.55.2:6152
     if [ ! -d ceph-ansible ]; then
         git clone https://github.com/ceph/ceph-ansible.git
     fi
     cd ceph-ansible/ && git checkout stable-5.0
-    pip install -r requirements.txt
+    pip3 install -r requirements.txt
 
     unset cp
     cp -f ${workdir}/ansible-config/site.yml .
